:building_construction: turn proxies.py into subpackage with individual resource modules
