+++ .pre-commit-config.yaml
@@ -26,7 +26,7 @@ repos:
     rev: v0.15.7
     hooks:
       - id: ruff # Fixes issues (including unused imports)
-        args: ["--fix"]
+        args: ["--fix", "--extend-ignore", "E402"]
         files: '^(mgnipy|tests)/.*\.py$'
         types: [python]
