+++ mgnipy/V2/__init__.py
@@ -5,14 +5,22 @@ from ..emgapi_v2_client.client import (
 from .core import MGnifier
 from .proxies import (
     Analyses,
+    Assemblies,
     Biomes,
+    Catalogues,
     Genomes,
+    Publications,
+    Runs,
     Samples,
     Studies,
 )
 
 __all__ = (
     "MGnifier",
+    "Assemblies",
+    "Catalogues",
+    "Publications",
+    "Runs",
     "Biomes",
     "Studies",
     "Samples",
