+++ mgnipy/V2/proxies/runs.py
@@ -1,0 +1,42 @@
+from typing import (
+    Any,
+    ClassVar,
+    Literal,
+    Optional,
+)
+
+from mgnipy.V2.proxies import MGnifyDetail, MGnifyList
+
+
+class Runs(MGnifyList):
+
+    RESOURCE: ClassVar[Literal["runs"]] = "runs"
+
+    def __init__(
+        self,
+        *,
+        params: Optional[dict[str, Any]] = None,
+        config: Optional[dict] = None,
+        **kwargs,
+    ):
+
+        super().__init__(params=params, config=config, **kwargs)
+
+
+class RunDetail(MGnifyDetail):
+    RESOURCE: ClassVar[Literal["run"]] = "run"
+
+    def __init__(
+        self,
+        id: Optional[str] = None,
+        *,
+        accession: Optional[str] = None,
+        config: Optional[dict] = None,
+        **kwargs,
+    ):
+
+        super().__init__(
+            id=id or accession,
+            config=config,
+            **kwargs,
+        )
