+++ mgnipy/V2/proxies/samples.py
@@ -1,0 +1,41 @@
+from typing import (
+    Any,
+    ClassVar,
+    Literal,
+    Optional,
+)
+
+from mgnipy.V2.proxies import MGnifyDetail, MGnifyList
+
+
+class Samples(MGnifyList):
+    RESOURCE: ClassVar[Literal["samples"]] = "samples"
+
+    def __init__(
+        self,
+        *,
+        params: Optional[dict[str, Any]] = None,
+        config: Optional[dict] = None,
+        **kwargs,
+    ):
+
+        super().__init__(params=params, config=config, **kwargs)
+
+
+class SampleDetail(MGnifyDetail):
+    RESOURCE: ClassVar[Literal["sample"]] = "sample"
+
+    def __init__(
+        self,
+        id: Optional[str] = None,
+        *,
+        accession: Optional[str] = None,
+        config: Optional[dict] = None,
+        **kwargs,
+    ):
+
+        super().__init__(
+            id=id or accession,
+            config=config,
+            **kwargs,
+        )
